Completely modularized all functions of the CLI
